# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenG2P Audit Manager",
    "description": "CloudEvents audit ingest for OpenG2P. Accepts events over HTTP, buffers in Kafka, persists to PostgreSQL. Callers receive 202 Accepted immediately \u2014 Kafka and Postgres are never on the hot path.",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/auditmanager/events": {
      "post": {
        "summary": "Ingest a single CloudEvent",
        "description": "Accepts one CloudEvent, validates its schema, and enqueues it for durable publication to Kafka.\n\nReturns **202 Accepted** the moment the event lands in the in-process queue \u2014 Kafka and Postgres latency are fully hidden from the caller. Delivery from there is at-least-once with idempotent inserts on (`id`, `occurred_at`), so replays after a crash never produce duplicate rows in the audit store.",
        "operationId": "ingest_event_v1_auditmanager_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Event accepted into the ingest queue. The `id` you submitted is echoed back.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Malformed CloudEvent \u2014 failed schema validation (missing required field, invalid `outcome` enum, unparseable `time`, etc.).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service not ready or backpressure.\n\n- `AUD-004` \u2014 ingest queue full. Retry with exponential backoff + jitter.\n- `AUD-005` \u2014 service startup not complete.\n- `AUD-006` \u2014 database health check failed (surfaces via `/health`, but also blocks ingest readiness indirectly).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auditmanager/events/batch": {
      "post": {
        "summary": "Ingest a batch of CloudEvents",
        "description": "Accepts up to `ingest.max_batch_size` CloudEvents in a single request. Each event is validated and enqueued independently \u2014 same 202 semantics as `/events`. Events in the batch are not atomic: on queue overflow part way through, the response reports the count actually accepted.\n\nEvents in a batch become **separate rows** in Postgres (one per CloudEvent). Batching is a transport optimization \u2014 not a storage concept.",
        "operationId": "ingest_batch_v1_auditmanager_events_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "All events (or as many as fit before backpressure) accepted. The response echoes accepted ids.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "`AUD-007` \u2014 batch payload exceeds the configured `ingest.max_batch_size`. Split into smaller batches.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Malformed CloudEvent \u2014 failed schema validation (missing required field, invalid `outcome` enum, unparseable `time`, etc.).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service not ready or backpressure.\n\n- `AUD-004` \u2014 ingest queue full. Retry with exponential backoff + jitter.\n- `AUD-005` \u2014 service startup not complete.\n- `AUD-006` \u2014 database health check failed (surfaces via `/health`, but also blocks ingest readiness indirectly).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedPayload": {
        "properties": {
          "accepted": {
            "type": "string",
            "title": "Accepted",
            "description": "The `id` of the CloudEvent that was accepted into the ingest queue.",
            "examples": [
              "01HXQ9R2V3K8Y2ZP7N5FJ6M4A1"
            ]
          }
        },
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "AcceptedPayload"
      },
      "AcceptedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Constant \u2014 identifies the service that produced this envelope.",
            "default": "openg2p.auditmanager",
            "examples": [
              "openg2p.auditmanager"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Envelope schema version (not the service version).",
            "default": "1.0",
            "examples": [
              "1.0"
            ]
          },
          "responsetime": {
            "type": "string",
            "title": "Responsetime",
            "description": "RFC3339 timestamp at which this response was produced.",
            "examples": [
              "2026-04-23T10:00:00.000Z"
            ]
          },
          "response": {
            "$ref": "#/components/schemas/AcceptedPayload"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "examples": [
              []
            ]
          }
        },
        "type": "object",
        "required": [
          "responsetime",
          "response"
        ],
        "title": "AcceptedResponse"
      },
      "Actor": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "user",
              "system",
              "service",
              "anonymous"
            ],
            "title": "Type",
            "default": "user"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Actor",
        "description": "Who (or what system) triggered the event."
      },
      "AuditData": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/Actor"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "failure",
              "denied"
            ],
            "title": "Outcome"
          },
          "resource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Resource"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "actor",
          "action",
          "outcome"
        ],
        "title": "AuditData",
        "description": "OpenG2P-specific `data` shape. Open \u2014 extra fields allowed per event type."
      },
      "BatchAcceptedPayload": {
        "properties": {
          "accepted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted",
            "description": "CloudEvent ids accepted into the ingest queue, in the order supplied.",
            "examples": [
              [
                "01HXQ9R2V...",
                "01HXQ9R2W...",
                "01HXQ9R2X..."
              ]
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of events accepted (matches `len(accepted)`).",
            "examples": [
              3
            ]
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "count"
        ],
        "title": "BatchAcceptedPayload"
      },
      "BatchAcceptedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Constant \u2014 identifies the service that produced this envelope.",
            "default": "openg2p.auditmanager",
            "examples": [
              "openg2p.auditmanager"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Envelope schema version (not the service version).",
            "default": "1.0",
            "examples": [
              "1.0"
            ]
          },
          "responsetime": {
            "type": "string",
            "title": "Responsetime",
            "description": "RFC3339 timestamp at which this response was produced.",
            "examples": [
              "2026-04-23T10:00:00.000Z"
            ]
          },
          "response": {
            "$ref": "#/components/schemas/BatchAcceptedPayload"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "examples": [
              []
            ]
          }
        },
        "type": "object",
        "required": [
          "responsetime",
          "response"
        ],
        "title": "BatchAcceptedResponse"
      },
      "CloudEvent": {
        "properties": {
          "specversion": {
            "type": "string",
            "const": "1.0",
            "title": "Specversion",
            "default": "1.0"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "datacontenttype": {
            "type": "string",
            "title": "Datacontenttype",
            "default": "application/json"
          },
          "traceparent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traceparent"
          },
          "data": {
            "$ref": "#/components/schemas/AuditData"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "source",
          "type",
          "data"
        ],
        "title": "CloudEvent",
        "description": "CloudEvents v1.0 envelope.\n\nSpec: https://github.com/cloudevents/spec/blob/v1.0.2/cloudevents/spec.md"
      },
...
      "ErrorDetail": {
        "properties": {
          "errorCode": {
            "type": "string",
            "title": "Errorcode",
            "description": "OpenG2P-assigned error code. Catalog: `AUD-004` (ingest queue full \u2014 backpressure), `AUD-005` (service not ready \u2014 startup incomplete), `AUD-006` (database health check failed), `AUD-007` (batch payload exceeds configured max size).",
            "examples": [
              "AUD-004"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message describing the failure.",
            "examples": [
              "Audit ingest queue full \u2014 backpressure"
            ]
          }
        },
        "type": "object",
        "required": [
          "errorCode",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "One error entry in the `errors[]` array of the response envelope."
      },
      "ErrorResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Constant \u2014 identifies the service that produced this envelope.",
            "default": "openg2p.auditmanager",
            "examples": [
              "openg2p.auditmanager"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Envelope schema version (not the service version).",
            "default": "1.0",
            "examples": [
              "1.0"
            ]
          },
          "responsetime": {
            "type": "string",
            "title": "Responsetime",
            "description": "RFC3339 timestamp at which this response was produced.",
            "examples": [
              "2026-04-23T10:00:00.000Z"
            ]
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response",
            "description": "Always `null` on an error response.",
            "examples": [
              null
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "description": "One entry per failure. Always at least one for a non-2xx response.",
            "examples": [
              [
                {
                  "errorCode": "AUD-004",
                  "message": "Audit ingest queue full \u2014 backpressure"
                }
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "responsetime",
          "errors"
        ],
        "title": "ErrorResponse",
        "description": "Shared shape for every non-2xx response \u2014 `response` is null, `errors` is populated."
      },
      "EventBatch": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/CloudEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "EventBatch",
        "description": "Batched ingest payload."
      },
...
      "Resource": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Resource",
        "description": "The primary object acted upon (optional for plain logins)."
      },
...
    }
  }
}